+++ backend/src/tools/stocks.py
